isp11: rockchip: v0.1.a
1. fix struct cifisp_hist_stat .hist_bins unsigned short to unsigned int;

+++ include/media/rk-isp11-config.h
@@ -154,7 +154,7 @@ struct cifisp_awb_stat {
 };
 
 struct cifisp_hist_stat {
-	unsigned short hist_bins[CIFISP_HIST_BIN_N_MAX];
+	unsigned int hist_bins[CIFISP_HIST_BIN_N_MAX];
 };
 
 /*! BLS mean measured values*/
